Move GraphQL-files to separate files
