+++ back-end/src/graphql/queries/Menus.ts
@@ -1,0 +1,10 @@
+import { GraphQLFloat, GraphQLInt, GraphQLObjectType } from 'graphql';
+
+export const MenuType = new GraphQLObjectType({
+    name: 'menu',
+    fields: () => ({
+        date: { type: GraphQLFloat },
+        menuId: { type: GraphQLInt },
+        recipeId: { type: GraphQLInt }
+    })
+});
